Breakout strategy succesfully extracted
