+++ app/strategies/ichimoku_cloud.py
@@ -1,0 +1,40 @@
+def calculate_base_line(coin_pair, unit):
+    """
+    Calculates (26 period high + 26 period low) / 2
+    Also known as the "Kijun-sen" line
+    """
+
+    closing_prices = get_closing_prices(coin_pair, 26, unit)
+    period_high = max(closing_prices)
+    period_low = min(closing_prices)
+    return (period_high + period_low) / 2
+
+def calculate_conversion_line(coin_pair, unit):
+    """
+    Calculates (9 period high + 9 period low) / 2
+    Also known as the "Tenkan-sen" line
+    """
+    closing_prices = get_closing_prices(coin_pair, 9, unit)
+    period_high = max(closing_prices)
+    period_low = min(closing_prices)
+    return (period_high + period_low) / 2
+
+def calculate_leading_span_a(coin_pair, unit):
+    """
+    Calculates (Conversion Line + Base Line) / 2
+    Also known as the "Senkou Span A" line
+    """
+
+    base_line = calculate_base_line(coin_pair, unit)
+    conversion_line = calculate_conversion_line(coin_pair, unit)
+    return (base_line + conversion_line) / 2
+
+def calculate_leading_span_b(coin_pair, unit):
+    """
+    Calculates (52 period high + 52 period low) / 2
+    Also known as the "Senkou Span B" line
+    """
+    closing_prices = get_closing_prices(coin_pair, 52, unit)
+    period_high = max(closing_prices)
+    period_low = min(closing_prices)
+    return (period_high + period_low) / 2
